Added key remapping
Well sort of, because the Teardown API doesn't provide a simple way of knowing which key is pressed by the user, I had to make a keyboard layout selector available through the mod options. There are two layout as of now, the QWERTY as a default and the AZERTY one, the user can select any one of them and the choice is stored in the save file of the mod using the Teardown registry.